problem100/p160_intersection_of_two_linked_lists.go
package problem100

/**
160. 相交链表
给你两个单链表的头节点 headA 和 headB ，请你找出并返回两个单链表相交的起始节点。如果两个链表不存在相交节点，返回 null 。
图示两个链表在节点 c1 开始相交：
题目数据 保证 整个链式结构中不存在环。
注意，函数返回结果后，链表必须 保持其原始结构 。
a1->a2->a3->a4
b1->a3->a4
return a3
 * Definition for singly-linked list.
 * type ListNode struct {
 *     Val int
 *     Next *ListNode
 * }
*/
func getIntersectionNode(headA, headB *ListNode) *ListNode {
	//pa:  a1->a2->a3->a4 -> b1->a3->a4
	//pb:  b1->a3->a4 -> a1->a2->a3->a4
	pa, pb := headA, headB
	for pa != pb {
		if pa == nil {
			pa = headB
		} else {
			pa = pa.Next
		}

		if pb == nil {
			pb = headA
		} else {
			pb = pb.Next
		}
	}
	return pa
}
